% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_funs.R
\name{get_WHO_region}
\alias{get_WHO_region}
\title{Look up a country's WHO region}
\usage{
get_WHO_region(this.country)
}
\arguments{
\item{this.country}{name of the input country (a string)}
}
\value{
Name of the corresponding WHO region
}
\description{
Look up a country's WHO region
}
\examples{
get_WHO_region("Germany")
get_WHO_region("China")
}
