% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_funs.R
\name{show_available_regions}
\alias{show_available_regions}
\title{Show all WHO regions}
\usage{
show_available_regions()
}
\value{
A data frame of valid region names.
}
\description{
Lists all WHO regions, with valid spelling and formatting.
}
\examples{
show_available_regions()
}
